proxy filtered-groups: Populate Region field API calls
