+++ docs/ecctl_platform_proxy_filtered-group_update.adoc
@@ -17,7 +17,7 @@ ecctl platform proxy filtered-group update <filtered group id> --filters <key1=v
 
 ----
       --expected-proxies-count int32   Expected proxies count in filtered group (default 1)
-      --filters stringToString         fFlters for proxies group (default [])
+      --filters stringToString         Filters for proxies group (default [])
   -h, --help                           help for update
       --version int                    Version update for filtered group
 ----
